+++ src/components/Header.tsx
@@ -1,296 +1,16 @@
-import Button from '@components/Button'
-import Icon from '@components/Icon'
-import PageLink from '@components/PageLink'
-import { NOTIFICATION_ICON_CONFIG, ROUTES, Z_INDEX } from '@constants/ui'
-import notificationsData from '@data/notificationsData'
-import type { NotificationItem } from '@src/types/notification'
-import {
-  Bell,
-  CalendarCheck,
-  Check,
-  UserRound,
-  UserRoundPlus,
-  UsersRound,
-  X,
-} from 'lucide-react'
-import { useEffect, useMemo, useRef, useState } from 'react'
-import { Link, useNavigate } from 'react-router'
+import { Z_INDEX } from '@constants/ui'
 
-const iconMap = {
-  UserRoundPlus,
-  Check,
-  X,
-  UsersRound,
-  CalendarCheck,
-}
-
-const getNotificationIcon = (type: NotificationItem['type']) => {
-  const config = NOTIFICATION_ICON_CONFIG.find((item) => item.type === type)
-
-  if (!config) {
-    return (
-      <div className="flex h-8 w-8 shrink-0 items-center justify-center rounded-full bg-gray-100">
-        <div className="h-5 w-5 rounded-full bg-gray-500" />
-      </div>
-    )
-  }
-  const IconComponent = iconMap[config.icon as keyof typeof iconMap]
-  // "config.icon은 iconMap의 키 중 하나야"라고 알려주는 것
-
-  return (
-    <div
-      className={`flex h-8 w-8 shrink-0 items-center justify-center rounded-full ${config.bgColor}`}
-    >
-      <Icon icon={IconComponent} className={config.strokeColor} size="s" />
-    </div>
-  )
-}
+import Logo from '@components/Logo'
+import Navigation from '@components/Navigation'
 
 const Header = () => {
-  const [notifications, setNotifications] =
-    useState<NotificationItem[]>(notificationsData)
-  const [isNotificationOpen, setIsNotificationOpen] = useState(false)
-  const [notificationFilter, setNotificationFilter] = useState<
-    'all' | 'unread' | 'read'
-  >('all')
-
-  const notificationRef = useRef<HTMLDivElement>(null)
-  const buttonRef = useRef<HTMLButtonElement>(null)
-  const navigate = useNavigate()
-
-  // 개발단계에선 해당 값을 true 와 false 로 합니다.
-  const user = true
-
-  const handleNotificationToggle = () => {
-    setIsNotificationOpen(!isNotificationOpen)
-  }
-
-  const filteredNotifications = useMemo(() => {
-    switch (notificationFilter) {
-      case 'unread':
-        return notifications.filter((notification) => notification.isUnread)
-      case 'read':
-        return notifications.filter((notification) => notification.isRead)
-      default:
-        return notifications
-    }
-  }, [notifications, notificationFilter])
-
-  const handleMarkAllAsRead = () => {
-    setNotifications((prev) =>
-      prev.map((notification) => ({
-        ...notification,
-        isRead: true,
-        isUnread: false,
-      }))
-    )
-  }
-
-  const linkToRecruitmentManage = () => {
-    navigate(ROUTES.RECRUITMENT_MANAGE)
-    setIsNotificationOpen(false)
-  }
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (
-        isNotificationOpen &&
-        notificationRef.current &&
-        buttonRef.current &&
-        !notificationRef.current.contains(event.target as Node) &&
-        !buttonRef.current.contains(event.target as Node)
-      ) {
-        setIsNotificationOpen(false)
-      }
-    }
-
-    document.addEventListener('mousedown', handleClickOutside)
-    return () => document.removeEventListener('mousedown', handleClickOutside)
-  }, [isNotificationOpen])
-
   return (
     <header
       className={`sticky top-0 z-[${Z_INDEX.HEADER}] flex h-16 w-full items-center justify-center px-20`}
     >
       <div className="flex h-full w-full max-w-7xl justify-between px-8">
-        <Link to="/" className="flex items-center gap-2">
-          <p className="bg-primary-500 flex h-8 w-8 items-center justify-center rounded-lg text-base font-bold text-white">
-            S
-          </p>
-          <p className="text-primary-600 text-xl font-bold">StudyHub</p>
-        </Link>
-        <nav className="flex items-center gap-8">
-          <Link
-            to={`/courses`}
-            className="flex items-center text-base text-gray-700"
-          >
-            강의 목록
-          </Link>
-          <Link
-            to={`/study-group`}
-            className="flex items-center text-base text-gray-700"
-          >
-            스터디 그룹
-          </Link>
-          <Link
-            to={`/recruitment`}
-            className="flex items-center text-base text-gray-700"
-          >
-            구인 공고
-          </Link>
-          <div className="flex items-center gap-4">
-            {user ? (
-              <>
-                <button
-                  ref={buttonRef}
-                  className="relative flex size-10 items-center justify-center rounded-full"
-                  onClick={handleNotificationToggle}
-                >
-                  <Icon icon={Bell} size="md" className={`stroke-gray-600`} />
-                  <p className="bg-danger-500 absolute -top-1 left-6 flex h-5 w-5 items-center justify-center rounded-full text-xs font-semibold text-white">
-                    3
-                  </p>
-                  {isNotificationOpen && (
-                    <div
-                      ref={notificationRef}
-                      className={`absolute top-12 right-0 z-[${Z_INDEX.DROPDOWN}] max-h-[819.2px] w-96 overflow-hidden rounded-lg border border-gray-200 bg-white`}
-                      onClick={(e) => e.stopPropagation()}
-                    >
-                      <div className="flex items-center justify-between px-4 pt-4 pb-[17px]">
-                        <h3 className="flex items-center justify-start text-lg font-semibold text-gray-900">
-                          알림
-                        </h3>
-                        <div
-                          onClick={handleMarkAllAsRead}
-                          className="text-primary-600 flex items-center justify-center text-center text-sm"
-                        >
-                          모두 읽음
-                        </div>
-                      </div>
-
-                      <div className="flex">
-                        <div
-                          role="tab"
-                          onClick={() => setNotificationFilter('all')}
-                          className={`flex w-2/3 items-center justify-center pt-3 pb-3.5 text-sm ${
-                            notificationFilter === 'all'
-                              ? 'border-primary-500 text-primary-600 border-[0px_0px_2px] border-solid'
-                              : 'border-[0px_0px_2px] border-solid border-transparent text-gray-500'
-                          }`}
-                        >
-                          전체보기 ({notifications.length})
-                        </div>
-                        <div
-                          role="tab"
-                          onClick={() => setNotificationFilter('unread')}
-                          className={`flex w-2/3 items-center justify-center pt-3 pb-3.5 text-sm ${
-                            notificationFilter === 'unread'
-                              ? 'border-primary-500 text-primary-600 border-[0px_0px_2px] border-solid'
-                              : 'border-[0px_0px_2px] border-solid border-transparent text-gray-500'
-                          }`}
-                        >
-                          읽지 않음 (
-                          {
-                            notifications.filter(
-                              (notification) => notification.isUnread
-                            ).length
-                          }
-                          )
-                        </div>
-                        <div
-                          role="tab"
-                          onClick={() => setNotificationFilter('read')}
-                          className={`flex w-2/3 items-center justify-center pt-3 pb-3.5 text-sm ${
-                            notificationFilter === 'read'
-                              ? 'border-primary-500 text-primary-600 border-[0px_0px_2px] border-solid'
-                              : 'border-[0px_0px_2px] border-solid border-transparent text-gray-500'
-                          }`}
-                        >
-                          읽음 (
-                          {
-                            notifications.filter(
-                              (notification) => notification.isRead
-                            ).length
-                          }
-                          )
-                        </div>
-                      </div>
-
-                      <div
-                        className="flex max-h-80 flex-col overflow-y-auto"
-                        role="tabpanel"
-                      >
-                        {filteredNotifications.map((notification) => (
-                          <div
-                            key={notification.id}
-                            className={`${notification.isUnread ? 'bg-primary-50' : 'bg-white'} flex flex-col content-stretch items-start justify-start border-gray-100`}
-                          >
-                            <div
-                              className="flex items-start gap-3 border-b border-gray-200 px-4 pt-[17px] pb-4"
-                              onClick={linkToRecruitmentManage}
-                            >
-                              {getNotificationIcon(notification.type)}
-                              {/* <div className="flex"> */}
-                              <div className="flex flex-col gap-1">
-                                {/* <div
-                                  className={`-webkit-box relative h-10 w-[286px] shrink-0 flex-col justify-center overflow-hidden font-['Pretendard:Regular',_sans-serif] text-[14px] leading-[0] overflow-ellipsis not-italic ${notification.isUnread ? 'text-gray-900' : 'text-gray-700'}`}
-                                  style={{
-                                    display: '-webkit-box',
-                                    WebkitLineClamp: 2,
-                                    WebkitBoxOrient: 'vertical',
-                                  }}
-                                > */}
-                                <p className="text-left text-sm text-gray-900">
-                                  {notification.message}
-                                </p>
-                                {/* </div> */}
-                                <p className="flex items-center text-xs text-gray-500">
-                                  {notification.date}
-                                </p>
-                              </div>
-                              {notification.isUnread && (
-                                <div className="flex pt-2">
-                                  <div className="bg-primary-500 size-2 rounded-full" />
-                                </div>
-                              )}
-                            </div>
-                          </div>
-                        ))}
-                      </div>
-                      <div className="flex h-[45px] bg-gray-50" />
-                    </div>
-                  )}
-                </button>
-                <a className="flex items-center gap-2">
-                  <div className="bg-primary-100 flex h-8 w-8 items-center justify-center rounded-full">
-                    <Icon
-                      icon={UserRound}
-                      size="sm"
-                      className={`stroke-primary-600`}
-                    />
-                  </div>
-                  <p className="text-base text-gray-700">김스터디</p>
-                </a>
-              </>
-            ) : (
-              <>
-                <PageLink
-                  pageLinkInnerText="로그인"
-                  variant="ghost"
-                  hoverTextColor="text-gray-700"
-                  fontWeight="normal"
-                  linkTo="/login"
-                />
-                <Button
-                  buttonInnerText="회원가입"
-                  size="base"
-                  bgColor="bg-primary-500"
-                  borderColor="border-primary-500"
-                  fontWeight="medium"
-                />
-              </>
-            )}
-          </div>
-        </nav>
+        <Logo />
+        <Navigation />
       </div>
     </header>
   )
